fraudlens_streamlit/requirements.txt
streamlit
pandas
scikit-learn
joblib
numpy
pillow
opencv-python
